Re-use childNodeCreator in the other Models.
This simple wrapper was introduced in the Community Model. I'm just using it in
the other models now where applicable.

`childNodeCreator` makes a function that acts like `createNode`, but adds in the
parent ID of the current node calls `createParentChildLink` between the two
nodes, and resolves when the prior two tasks do.

I invoke the creators in the "parent" models' APIs so the Models that consume
them get a nice function for creating child nodes. The consumer API calls that
use this helper don't even need to destructure `actions` anymore, as making a
child node is their only function most of the time.
