+++ src/gatsby/models/docs/onCreateMarkdownContentNode.js
@@ -1,6 +1,6 @@
 const path = require('path')
 
-function createMarkdownDocsNode(api, { parentNode }) {
+async function createMarkdownDocsNode(api, { parentNode, createChildNode }) {
   // Suppress page creation for Basic Concepts and the Glossary
   // They're only used in tooltips now, but we intend to expand on them later.
   if (parentNode.relativeDirectory === 'docs/user-guide/basic-concepts') return
@@ -8,8 +8,7 @@ function createMarkdownDocsNode(api, { parentNode }) {
   const splitDir = parentNode.relativeDirectory.split('/')
   if (splitDir[0] !== 'docs') return
 
-  const { node, actions, createNodeId, createContentDigest } = api
-  const { createNode, createParentChildLink } = actions
+  const { node, createNodeId, createContentDigest } = api
   const { name, relativePath } = parentNode
   splitDir[0] = 'doc'
 
@@ -33,8 +32,7 @@ function createMarkdownDocsNode(api, { parentNode }) {
     }
   }
 
-  createNode(docNode)
-  createParentChildLink({ parent: node, child: docNode })
+  return createChildNode(docNode)
 }
 
 module.exports = createMarkdownDocsNode
