+++ src/gatsby/models/json-files/index.js
@@ -1,14 +1,17 @@
 const callOnModels = require('../../utils/models')
+const { childNodeCreator } = require('../../utils/models/nodes.js')
 
 module.exports = {
   async onCreateNode(api, { models }) {
     const { node, loadNodeContent } = api
+    const createChildNode = childNodeCreator(api)
     if (node.internal.type !== 'File' || node.extension !== 'json') return null
 
     const content = JSON.parse(await loadNodeContent(node))
 
     return callOnModels(models, 'onParseJsonFile', api, {
-      content
+      content,
+      createChildNode
     })
   }
 }
